refactor(client): change folder structure
