+++ packages/client/src/hooks/queries/comment/useLikeComment.ts
@@ -1,4 +1,4 @@
-import CommentAPI from '~/api/comment';
+import CommentAPI from '~/lib/api/comment';
 import { useMutation } from '@tanstack/react-query';
 import { UseMutationOptionsOf } from '~/hooks/queries/types';
