+++ packages/client/src/hooks/usePostLikeManager.ts
@@ -1,6 +1,6 @@
 import { useQueryClient } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
-import { PostStats, PostWithStats } from '~/types';
+import { PostStats, PostWithStats } from '~/lib/types';
 import { useGetPost } from './queries/post';
 import useLikePost from './queries/post/useLikePost';
 import useUnlikePost from './queries/post/useUnlikePost';
