+++ packages/client/src/hooks/useUser.ts
@@ -1,5 +1,5 @@
 import { useQueryClient } from '@tanstack/react-query';
-import { User } from '~/types';
+import { User } from '~/lib/types';
 import { useGetMe } from './queries/user';
 
 const useUser = () => {
